microservices/product_service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
redis>=4.5.0
django-redis>=5.2.0
Pillow>=9.0.0
gunicorn>=20.1.0
django-cors-headers>=3.13.0
boto3>=1.26.0
django-storages>=1.13.0